requirements.txt
streamlit
openai-whisper
torch
groq
python-dotenv
